+++ lib/build_eval/server/travis.rb
@@ -8,12 +8,12 @@ module BuildEval
       end
 
       def build_result(name)
-        response = BuildEval::Http.get(
-          "https://api.travis-ci.org/repositories/#{@username}/#{name}/cc.xml"
-        )
+        response = BuildEval::Http.get("https://api.travis-ci.org/repositories/#{@username}/#{name}/cc.xml")
         build_element = Nokogiri::XML(response.body).xpath("//Project").first
         raise "Unexpected build response: #{response.message}" unless build_element
-        BuildEval::Result::BuildResult.create(build_name: name, status_name: build_element.attribute("lastBuildStatus").value)
+        BuildEval::Result::BuildResult.create(
+          build_name: name, status_name: build_element.attribute("lastBuildStatus").value
+        )
       end
 
       def to_s